+++ RELEASE_NOTES.md
@@ -27,6 +27,8 @@ RELEASE NOTES
 Documentation
 -------------
 
+* ADDED: Library notes to the Handbook.
+
 * ADDED: First version of a Handbook section on Python nomenclature.
 
 * IMPROVED: Handbook section on Java nomenclature.